Azure/requirements.txt
tensorflow==1.14.0
azureml-sdk==1.0.65
azureml-widgets==1.0.65